Moved logout/regenerate token functionality from broadcast receiver in handler classes.
