+++ app/src/main/java/com/koleff/kare_android/common/broadcast/RegenerateTokenNotifier.kt
@@ -1,8 +1,10 @@
-package com.koleff.kare_android.common.network
+package com.koleff.kare_android.common.broadcast
 
 import com.koleff.kare_android.ui.state.TokenState
 import kotlinx.coroutines.flow.Flow
 
 interface RegenerateTokenNotifier {
     val regenerateTokenState: Flow<TokenState>
+
+    //suspend fun notifyTokenRegenerationResult(result: ResultWrapper<Boolean>)
 }
